+++ bricks/__init__.py
@@ -2,6 +2,6 @@
 
 from bricks.api import Bricks
 
-__version__ = "0.4.18"
+__version__ = "0.4.19"
 
 __all__ = ["Bricks", "__version__"]